Update patch set 4
Patch Set 4:

(1 comment)

This approach seems OK to me.  It's hard to believe that people are using this mux with millions of entries.  And, as you say, if they were, they were already paying a penalty in the more frequent lookup case.

That said I don't understand why we do this linear search at all.  As I understand it we are only looking for entries that end with a '/'.  Why not something like (completely untested)
    r := strings.LastIndexByte(path, '/')
    for r != -1 {
        if v, ok := mux.m([path[:r+1]); ok {
            return v.h, v.pattern
        }
        r = strings.LastIndexByte(path[:r], '/')
    }

Patch-set: 4
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>