runtime: look for idle p to run current thread when quantum expires into GC
This repairs one of the several causes of pauses uncovered
by a GC microbenchmark.  When a goroutine's quantum expires,
if a GC is needed, that means that the amount of available
work has expanded by one.  If the runtime does not check for
an idle P, it will leave the previously-current goroutine
not running even when capacity to run it exists.  The fix
is to do that check.

This is problem 4 in this list:
https://github.com/golang/go/issues/27732#issuecomment-423301252

Updates #27732.

Change-Id: I6905365cac8504cde6faab2420f4421536551f0b