Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > I'm not sure the internal/bytealg changes are correct. Or at least make sense.
> > > The assembly in internal/bytealg/compare_amd64.s, for example, has:
> > > 
> > > TEXT ·Compare(SB),NOSPLIT,$0-56
> > > TEXT bytes·Compare(SB),NOSPLIT,$0-56
> > > TEXT runtime·cmpstring(SB),NOSPLIT,$0-40
> > > 
> > > If someone called bytes.Compare in user code, before this CL it goes to that second one.  With your alias
> > > 
> > > //go:linkname bytes_Compare bytes.Compare
> > > func bytes_Compare(a, b []byte) int
> > > 
> > > Then it would go to internal/bytealg.bytes_Compare?  There's no implementation of that function anywhere.
> > 
> > I'm not sure I'm following. Why would it go to internal/bytealg.bytes_Compare? Packages that import bytes have no idea about internal/bytealg.bytes_Compare. That's just a local Go symbol for the link symbol "bytes.Compare".
> 
> So I don't understand what this CL does then.  It adds another Go definition for the same assembly code?  Then there's one in bytes/bytes_decl.go and one in internal/bytealg/compare_native.go? They have different names but both linkname to the same assembly implementation?
> It all seems terribly confusing. Especially if the local name might as well be internal/bytealg.foo.

In order to generate an ABI wrapper for Go code to call the assembly implementation of bytes.Compare, we need the definition and the Go stub in the same package. Having the assembly definition here tells the compiler that it's ABI0 (the stable/stack ABI) and the Go stub gives it the type information necessary to generate the internal ABI wrapper. (Of course, right now stable and internal ABIs are identical, so we're not using the type information, but we presumably will want to once they diverge.)

Patch-set: 1
Reviewer: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>