Update patch set 5
Patch Set 5:

> The change looks right to me but it would be nice to have a test somehow.

I've added a test and verified locally (for what little that's worth).

Expanding on this: There are two outcomes in the "done" mode (that Wil Selwood added) that are relevant to chatty mode. The first is that the output buffer does not receive the log line (because fmt.Printf printed it instead), and the second is that a subtest's log line is not lost.

I've added a test for the former, but I couldn't think of a way to add a test for the latter without adding a level of indirection between the logDepth and fmt.Printf (which the test would introspect).

Do you have any advice? Should I leave it as is, or should we add more scaffolding to make that assertion?

> Patch Set 4:
> 
> Bump, Jean? If we want this in 1.12, it should be merged very soon.

Sorry for the slow reply! (lots of traveling)

Patch-set: 5
Reviewer: Gerrit User 26615 <26615@62eb7196-b449-3ce5-99f1-c037f21e1705>