Update patch set 3
Patch Set 3:

(8 comments)

We should really be using #defines for the registers used in the new functions. Especially for the arguments. That is the convention used in the rest of the code.

Another thing that would be good is making the *Sqr* functions take one argument rather than two. That would need an extra stub for the VX code but would make it clearer that we are squaring.

Patch-set: 3
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>