Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> That test failure is a timeout in the new misc/cgo test that I added. The good news is that I can reproduce locally without the changes to cgocall.go, so it would appear to be a bug in the test and not in the cgocall.go changes.
> 
> I'm not entirely sure what's going wrong; the test just sometimes wedges for 60s. It would seem that the sleeper goroutines are just starving the main test goroutine. I wonder if #10958 is relevant. The sleeper goroutines might not have a preemption point because they call into cgo in a tight loop. (And indeed, sticking a runtime.Gosched() call into that loop fixes the problem.) But I wonder if something else is going wrong. You'd expect P-retaking to allow the other goroutines to make progress.

If the theory is that it's not a new bug, then I would not block this on that. Just add a Gosched, a comment and a reference to the failure and maybe file a new bug, or update the old one.

Patch-set: 3
Reviewer: Gerrit User 5400 <5400@62eb7196-b449-3ce5-99f1-c037f21e1705>