runtime: preempt goroutines that make repeated syscalls/cgo calls
Goroutines that made repeated cgo calls or syscalls could starve out
other goroutines. The problem was especially acute when the calls were
long enough and frequent enough that sysmon would be unlikely to observe
the goroutine in-between the syscalls/cgo calls, but short enough that
sysmon would never observe the goroutine in the same syscall/cgo call
twice. (The magic number is about 1ms.) Since no single call was long,
sysmon would never consider the P for retaking, and since the status of
the P was _Psyscall and not _Prunning, sysmon would never consider the
goroutine for preemption.

The solution is simply to teach sysmon to consider goroutines for
preemption when their P is in either _Psyscall or _Prunning.

Fix #28701.

Change-Id: I0b0c196f9d4075be2a3213717c28e93baf006ba3