Update patch set 1
Patch Set 1:

(3 comments)

I appreciate the effort to add examples to more and more
functions in the stdlib. But I think great care should be
taken to add examples which really illustrate not only
syntax and trivialities but the "non-obvious", the non-trivial
aspects of these functions. Maybe you can improve these
examples in that dimension (because math.Min(2,3) feels just
too trivial).

Patch-set: 1
CC: Gerrit User 5050 <5050@62eb7196-b449-3ce5-99f1-c037f21e1705>