Update patch set 1
Patch Set 1:

> The issue required the explanation and a possible workaround, what would be the best possible place to put them?

It may be appropriate to add something to the package docs.  Or better might be to add an example.

Note that we don't have to do exactly what the issue says, we have to find the right fix.

Thanks.

Patch-set: 1
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>