Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What are "some implementations"?

msys2 (uname -a MSYS_NT-6.3 skinny 2.10.0(0.325/5/3) 2018-02-09 15:25 x86_64 Msys) is just one example which can declare errno in sys/errno.h as:

#ifndef _REENT_ONLY
#define errno (*__errno())
extern int *__errno (void);
#endif

But one thing which is weird is that "errno.h" is included in all the other stubs except for the 3 in this patch (for some reason). Wasn't sure why it was skipped for these because it's essentially not declaring errno.

Patch-set: 1
Reviewer: Gerrit User 27538 <27538@62eb7196-b449-3ce5-99f1-c037f21e1705>