Update patch set 2
Patch Set 2:

FWIW, if you see code going out of its way to do something odd like this, a good question to ask is whether it is out of sync with the surrounding code, and if so, which is wrong. In this case the print statement was correct but the computation of prefix was not. I sent CL 149317.

Patch-set: 2
CC: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>