Update patch set 2
Patch Set 2:

I am sorry but I am going to revert this change. I wrote the original code that way because I found it easier to read when the +1 and -1 results are symmetric. The code is perfectly valid as is.

In general golint makes suggestions. They don't need to be followed if they make the code worse.

Patch-set: 2
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>