Update patch set 2
Patch Set 2:

agree that it seems unlikely to trigger. was only sharing the first few optimal cases from my generator program should the concept take root. I do not know of any fp issues with this...think of it as associativity and the progressive terms as common subexpressions representing parenthesized factors. Knuth remarks on the idea and gives a few solutions so I just wrote my own in response to the need for quick integer powering.

this is not an argument for or against this submission, of which i know nothing.

Michael T. Jones

Patch-set: 2
Tag: mailMessageId=<CALoEmQyGUos7WcAxPTUQMwXY2_7EctOdwLn3r2a7MxSN0rpjUQ@mail.gmail.com>