cmd/trace: revert internal/traceparser
The performance improvement is not as big as we hoped.
Until the API is feature complete, we postpone the release
and avoid added complexity.

Revert "cmd/trace: use new traceparser to parse the raw trace files"
This reverts commit 08816cb8d7ed16b9c804587ff02c1ad1c3af6cd5.

Revert "internal/traceparser: provide parser that uses less space and parses segments of runtime trace files"
This reverts commit daaf361f74c3665bcb364356c5a9dd9f536c78c3.

Change-Id: Ic2a068a7dbaf4053cd9674ca7bde9c58e74385b4