Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > I am a bit wary of adding new API like this. We can already achieve this by manually calling the "then" function ourselves, which the net/http library already does.
> > > 
> > > This is just syntactic sugar and I am afraid might add precedent for adding more helpers like this.
> > > 
> > > I think this should be in a separate library and not inside "syscall/js".
> > 
> > We already have precedent: https://github.com/golang/go/issues/28264
> > 
> > This is one of the basic JavaScript operators. If we don't add it, then I expect people to ask for it a lot, because they expect to be able to translate their code 1-to-1 from JavaScript to Go.
> 
> I also expect this to be a quite common pattern, even more common than Truthy().

To me, syscall/js should just contain the bare minimum and nothing else. Truthy() was ok because it was hard to implement that in one's own code. (Considering that you will have to look up the spec and implement all that).

Compared to that, this seems easily possible to build on top of existing API.

Patch-set: 1