Update patch set 1
Patch Set 1:

It is not yet decided if we want to merge this or not. My initial reasoning was to add Await because it corresponds to a JavaScript operator, just for the ease of translating existing JavaScript code to Go. However, the additional useful functionality that you are pointing out shows that this is a more complex issue. I don't think that it would be good to add everything to syscall/js. So maybe we should instead consider to add an official "promise" package to one of the other Go subrepos?

Patch-set: 1
Reviewer: Gerrit User 13620 <13620@62eb7196-b449-3ce5-99f1-c037f21e1705>