Update patch set 1
Patch Set 1:

Thanks for putting up the CL as a starting point for discussion: it's an incredibly good way to make the conversation concrete!

> I am a bit wary of adding new API like this. We can already achieve this by manually calling the "then" function ourselves, which the net/http library already does.
> 
> This is just syntactic sugar and I am afraid might add precedent for adding more helpers like this.

I'm in the same boat: I think this is API creep. And I (now) feel the same about https://go-review.googlesource.com/c/go/+/144384/, particularly if we start citing that as precedent for changes like this.

syscall/js should, to my mind, represent the fundamentals of JavaScript. Aren't we effectively targeting ES5 VMs with syscall/js? If so, like Agniva said, I think this represents syntactic sugar. 

The entire of the Await function is something I could have written myself, outside of syscall/js. That's not to say there isn't a place for another x/tools (?) package that provides the next "layer" (with whatever feature-du-jour is most popular), I just don't think we should bloat syscall/js.

"Hiding" a blocking operation in this function also feels like it will trip people up (I'm not sure whether https://github.com/golang/go/issues/28911#issuecomment-441071266 is in fact the first instance of such a gotcha).

Patch-set: 1
CC: Gerrit User 16375 <16375@62eb7196-b449-3ce5-99f1-c037f21e1705>