Update patch set 1
Patch Set 1:

(1 comment)

Thank you for making this change Alex!
I've added a suggestion for the commit message and I'll
also tag Ian as a reviewer.

However, I am not sure if this change is safe to go in during
the cycle -- it seems harmless but I'll defer to Ian.

Patch-set: 1
CC: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>