net/http: close connection if `Connection: close` is set
As https://tools.ietf.org/html/rfc2616#page-117 says:

HTTP/1.1 defines the "close" connection option for the sender to signal that the connection will be closed after completion of the response.

But it does not specify the connection should be closed by client or server, for safety, if `Connection: close`
is set in response header, server should close the connection after write data. Recently we have been
affect by this issue: we write a lots of data to client, but the client does not close the connection(
keep-alive is set by default in HTTP/1.1), so memory is still hold by `response`, and memory is used up.

Change-Id: I756f81ec85ab5667be8b4cf45625e53a8f7dd36c
GitHub-Last-Rev: e956b9cce8346928d7e0434511a99995b979266e
GitHub-Pull-Request: golang/go#28936