runtime: support for getting snapshots from perf record
DO NOT REVIEW.

This makes it possible to get snapshots of the Intel PT ring buffer
from within the runtime in cooperation with "perf record". This is
really useful for debugging; for example, you can see the last many
thousands of instructions that were executed before reaching a bad
condition. It's also potentially useful for performance analysis,
since Intel PT traces can be timestampped.

Change-Id: I6003e65510b59d0468b547b7f2c27e519574e07d