runtime: avoid holding the heap lock for GC pressure callback write barrier
There exists the potential for self-deadlock during the assignment of
mheap_.gcPressureChange which is done with the heap lock held. This
assignment may invoke the write barrier, which may result in a call to
allocManual, which also acquires the heap lock.

See #28993 for more details.

This change moves gcPressureChange out of mheap and instead makes it
its own global, such that it's still considered a GC root. It then
protects it with its own mutex, rather than relying on the mheap mutex.
This change also does the same for maxHeap, because maxHeap and
gcPressureChange must be kept in sync, and should thus be protected by
the same lock.

Change-Id: Ia0cb5c5c9ce9708adbf545529368f732f304d22f