Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> Seems to break s390x build

This looks like a problem with the S390X rules.

ADDconst requires an int32.
The Move rewrite does:

(Move [s] dst src mem) && s > 1024 ->
	(LoweredMove [s%256] dst src (ADDconst <src.Type> src [(s/256)*256]) mem)

For a move of size 1<<31, that generates a ADDconst with size 1<<31, which doesn't fit in an int32.

Not really sure why this CL causes us to trip on this bug, but that bug is there.

So, can ADDconst take an int64? or a uint32? Or do we need to use (ADD (MOVDconst ...))?
Moves are pretty rare, we can just use ADD+MOVDconst and the ADD+MOVDconst->ADDconst rule will clean it up if it can.

Patch-set: 6