cmd/go: emit go list error for local non-existant packages
In https://go-review.googlesource.com/129061, a check was added for
patterns that reference non-existent local directories. While this
prevented unnecessary network lookups (fixing #26874), it caused
"go list -e" to exit with an error instead of listing packages with
error messages.

This change avoids the network lookup and does not exit for these
kinds of packages. Errors are still reported by
internal/load.LoadImport for packages that don't exist.

Fixes #28023

Change-Id: I0a648269e437aed3a95bfb05461a397264f3793f