Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > ping Robert.
> > > 
> > > I would need a bit of your guidance on this. Please see my comment on the last changeset. Specifically this one - https://go-review.googlesource.com/c/go/+/141617/2/src/go/doc/reader.go#418.
> > > 
> > > Thank you.
> > 
> > Sorry, but this will have to wait a bit. Since this is work in progress and fixes a bug, we will get this in for 1.12. But there are higher-priority issues at the moment.
> 
> Robert,
> 
> Just wanted to check on this since the beta1 release is around the corner. Would you be able to take a look sometime ? 
> 
> If you think this can go after beta releases, that is also fine.

We should get this in for the Beta, ideally.

Regarding the heuristic:

- iterate through the result list
- ignore predeclared types (aliasing to a predeclared type is a corner case)
- if we have more than one local type, do not associate

baseTypeName tells you if a name is imported (not local) or not.
We can't easily in this code tell if a built-in got redeclared locally - we'd have to know the whole package first and do the association later. Not hard, but not for this CL.

Patch-set: 3
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>