Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > That does mean that module owners can't play games with “required replacements” and still maintain compatibility with `go get`, but that's probably a good thing: we don't want to encourage replacement games in the first place. (Modules should always be able to build using their stated requirements!)
> 
> I'm concerned about this approach.
> 
> It means that exclude and replace statements have essentially no weight at all to anyone except someone that's directly developing within that module. Given that, I'd question whether they actually belong in the `go.mod` file at all. If a top level binary has no control over its exclusions and replacements, then why do we have them other than for local development?
> 
> Yes, a command _should_ be able to work with its stated requirements, but that might not be possible in some cases. Sometimes these "games" are actually essential, given that people do inadvertently make breaking changes.
> 
> For example, say I maintain a command A that depends on B and C, and B also depends on C. An important security update is issued to B which also incidentally bumps the version of C to a version that doesn't work with A (perhaps it builds but does not work for some reason). If `go get` respects exclude/replace statements in A, I can fix that and `go get A` will produce a working command containing the security update.

`go get A` will use exactly the dependencies that `A` was published with. (Remember, modules are designed to produce reproducible builds!) If you want to pick up patch releases you have to run `go get -u=patch A`, in which case you should also be able to combine a specific (fixed or earlier-than-broken) version of the dependency:

    go get -u=patch A C@version-that-fixes-regression

or

    go get -u=patch A B@version-that-works-with-c

You only need to apply a replacement if the owners of B and C refuse to fix the breakage. In that case (which should be very rare!), you can always create a temporary module in which you can specify the replacements explicitly and complete the build.

Patch-set: 10