Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 3:
> > > > 
> > > > > Patch Set 3:
> > > > > 
> > > > > > Patch Set 3:
> > > > > > 
> > > > > > ping Robert.
> > > > > > 
> > > > > > I would need a bit of your guidance on this. Please see my comment on the last changeset. Specifically this one - https://go-review.googlesource.com/c/go/+/141617/2/src/go/doc/reader.go#418.
> > > > > > 
> > > > > > Thank you.
> > > > > 
> > > > > Sorry, but this will have to wait a bit. Since this is work in progress and fixes a bug, we will get this in for 1.12. But there are higher-priority issues at the moment.
> > > > 
> > > > Robert,
> > > > 
> > > > Just wanted to check on this since the beta1 release is around the corner. Would you be able to take a look sometime ? 
> > > > 
> > > > If you think this can go after beta releases, that is also fine.
> > > 
> > > We should get this in for the Beta, ideally.
> > > 
> > > Regarding the heuristic:
> > > 
> > > - iterate through the result list
> > > - ignore predeclared types (aliasing to a predeclared type is a corner case)
> > > - if we have more than one local type, do not associate
> > > 
> > > baseTypeName tells you if a name is imported (not local) or not.
> > > We can't easily in this code tell if a built-in got redeclared locally - we'd have to know the whole package first and do the association later. Not hard, but not for this CL.
> > 
> > 
> > This is fine, except it breaks testdata/b.go, which has the "aliasing to a predeclared type" case
> > 
> > testdata/b.go:
> > ```
> > // Should only appear if AllDecls is set.
> > type uint struct{} // overrides a predeclared type uint
> > ```
> > 
> > Hence my current logic, which allows a predeclared type for i==0 and performs some extra bookkeeping to allow for signatures like "func F9() (int, T1, T2) {" to be associated.
> > 
> > What do you think we should do ?
> 
> What if we don't ignore predeclared types in step 2 if AllDecls is set? Not ideal, but might get us through the test cases.

Yeah, that fixes the testdata/b.go test case. But it breaks issue12839.1.golden, i.e. the case with AllDecls set and params like func F7() (T1, string). Since we are not ignoring predeclared types when AllDecls is set, this does not get matched with T1.

> 
> There's also cleanupTypes, maybe that can be used to fix things after the fact.

cleanupTypes is already called in doc.New. And from what I see, it removes the association, not create it. We need to make the association, since not ignoring predeclared types when AllDecls is set breaks the association.

If we allow predeclared types for i==0, that makes the testdata/b.go pass and also the issue12839.1.golden case. But that only breaks the case of "func F8() (int, T1, string)" in issue12839.1.golden, since there is an int at index 0 (one might even argue whether F8 should indeed be associated with T1 or not). Otherwise, this allows for most common cases and fixes the problem in the issue.

We can mark this case as a TODO and revisit this when we can catch the "built-in got redeclared locally" case.

Alternatively, maybe we can just hardcode to n == "error" and keep it at that. That doesn't need AllDecls check or i==0 check. Just fails cases like func F7() (T1, string). This is a big hammer of course, but still fixes the problem in the issue :)

I don't see any other less invasive solution at this point. Let me know your thoughts.

Patch-set: 3
Reviewer: Gerrit User 24096 <24096@62eb7196-b449-3ce5-99f1-c037f21e1705>